[UniversalRP] Changing some normal logic to use floats to avoid artifacts
Backport of https://jira.unity3d.com/browse/UUM-11468. Simple changes that make some halfs to floats to avoid banding artifacts.

+++ Packages/com.unity.render-pipelines.universal/ShaderLibrary/Input.hlsl
@@ -36,7 +36,7 @@ struct InputData
 {
     float3  positionWS;
     float4  positionCS;
-    half3   normalWS;
+    float3   normalWS;
     half3   viewDirectionWS;
     float4  shadowCoord;
     half    fogCoord;
